requirements.txt
flask
pandas
google-generativeai